BUG: Fix finding of JsonCpp_INCLUDE_DIRS.
Since the upstream JsonCpp does not yet have a jsoncpp-config.cmake, it has
difficulty finding the the include directory and libraries. The
FindJsonCpp.cmake module would fail with a system 0.5 JsonCpp installed
because it would find the wrong version when a build tree version was
intended.

We make sure the hint derived from JsonCpp_DIR looks in the source directory
instead of the build directory.